+++ src/main/scala/csv/instances/RegionIdInstances.scala
@@ -5,9 +5,10 @@ package instances
 import fs2.data.csv.*
 import at.ac.oeaw.imba.gerlich.gerlib.imaging.ImagingTimepoint
 
+/** CSV-related typeclass instances for region IDs */
 trait RegionIdInstances:
+    /** Encode the region ID by encoding simply the underlying imaging timepoint. */
     given cellEncoderForRegionId(
         using encTime: CellEncoder[ImagingTimepoint]
     ): CellEncoder[RegionId] = encTime.contramap(_.get)
-
 end RegionIdInstances