+++ src/main/scala/csv/instances/geometry/package.scala
@@ -1,0 +1,5 @@
+package at.ac.oeaw.imba.gerlich.looptrace.csv
+package instances
+
+/** CSV-related typeclass instances for geometry-related data types */
+package object geometry extends GeometryInstances