+++ src/main/scala/csv/instances/traceId/package.scala
@@ -1,0 +1,4 @@
+package at.ac.oeaw.imba.gerlich.looptrace.csv
+package instances
+
+package object traceId extends TraceIdInstances